requirements.txt
Django>=4.2,<5.0
pdfplumber>=0.10.0
python-docx>=1.0.0
spacy>=3.7.0
nltk>=3.8.1
click>=8.1.7
typer>=0.9.0
# If your project uses Pillow for ImageField/FileField images
Pillow>=10.0.0

# Optionally include these for user auth and file uploads
django-crispy-forms>=2.0
django-widget-tweaks>=1.4.12

# If you use PostgreSQL:
psycopg2>=2.9.6

# If using pip-tools for locking
# pip-tools>=7.3.0